+++ tests/dummy/app/users/user/controller.js
@@ -5,6 +5,9 @@ export default Controller.extend({
   folderExists: false,
   session: service('session'),
   folderService: service('folders-service'),
+  userService: service('user-service'),
+  portalService: service('portal-service'),
+  notificationType: 'email',
   actions: {
     checkFolderExists (folderName) {
       let username = this.get('session.currentUser.username');
@@ -37,6 +40,35 @@ export default Controller.extend({
         // remove it from the model list...
         this.get('model.folders').without(folder);
       });
+    },
+    removeNotification (notification) {
+      this.get('userService').removeNotification(notification.id)
+      .then(resp => {
+        this.send('forceRefresh');
+      })
+      .catch(_ => { alert('boo'); });
+    },
+    sendNotification (subject, message, notificationType) {
+      const methodNameHash = {
+        email: 'sendEmailNotification',
+        builtin: 'sendBuiltinNotification',
+        push: 'sendPushNotification'
+      };
+      const methodName = methodNameHash[notificationType];
+
+      const args = [message, [this.get('session.currentUser.username')]];
+      if (notificationType === 'push') {
+        // TODO: add clientId!
+      } else {
+        args.unshift(subject);
+      }
+
+      this.get('portalService')[methodName](...args)
+      .then(resp => {
+        this.setProperties({ notificationSubject: undefined, notificationMessage: undefined, notificationType: 'email' });
+        this.send('forceRefresh');
+      })
+      .catch(_ => alert('boo'));
     }
   }
 });