+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### added
+- `user-service` `getNotifications` and `removeNotification` methods
+- `portal-service` `sendEmailNotification`, `sendBuiltinNotification`, & `sendPushNotification` methods
+- `groups-service` `sendEmailNotification`, `sendBuiltinNotification`, & `sendPushNotification` methods which use AGRjs `createGroupNotification`
+
+### changed
+- upgraded arcgis-rest-js to 1.9.0
+- deprecated `portal-service` `sendMessage` method
+- deprecated `groups-service` `sendGroupMessage` method
 - configure eslint to use default ember rules & avoid jQuery
 
 ## [1.7.1]